Webpack 4 with ts-loader
