--- scripts/app/configs/webpack/common/node.ts
@@ -1,15 +1,0 @@
-import {Node} from "webpack";
-
-/**
- * @description Some libraries import Node modules but don"t use them in the browser.
- * Tell Webpack to provide empty mocks for them so importing them works.
- */
-const node: Node = {
-    child_process: "empty",
-    dgram: "empty",
-    fs: "empty",
-    net: "empty",
-    tls: "empty",
-};
-
-export default node;
