--- scripts/app/configs/webpack/production/server/module.ts
@@ -1,8 +1,0 @@
-import {typeScriptRule} from "../../common/rules";
-
-const module = {
-    rules: [typeScriptRule],
-    strictExportPresence: true,
-};
-
-export default module;
