--- scripts/app/configs/webpack/production/server/output.ts
@@ -1,10 +1,0 @@
-import Application from "../../../../Application";
-import files from "../../../files";
-
-const output = {
-    filename: files.outputPattern.jsServer,
-    path: Application.buildPath,
-    publicPath: Application.servedPath,
-};
-
-export default output;
