+++ tsconfig.json
@@ -13,7 +13,11 @@
     "allowJs": false,
     "jsx": "react",
     "moduleResolution": "node",
-    "rootDir": ".",
+    "rootDirs": [
+      "./.storybook",
+      "./scripts",
+      "./src"
+    ],
     "strict": true,
     "forceConsistentCasingInFileNames": true,
     "noImplicitReturns": true,
@@ -22,32 +26,15 @@
     "strictNullChecks": true,
     "suppressImplicitAnyIndexErrors": true,
     "noUnusedLocals": true,
-    "declaration": true,
+    "declaration": false,
     "esModuleInterop": true,
     "allowSyntheticDefaultImports": true
   },
-  "files": [
-    "./.storybook/addons.ts",
-    "./.storybook/config.ts",
-    "./.storybook/webpack.config.ts"
-  ],
   "typeRoots": [
     "./definitions",
     "./node_modules/@types"
   ],
   "types": [
     "jest"
-  ],
-  "include": [
-    "./.storybook",
-    "./definitions",
-    "./scripts",
-    "./src"
-  ],
-  "exclude": [
-    "build",
-    "coverage",
-    "node_modules",
-    "public"
   ]
 }
